file structure changed+
changed file structure so hopefully its more meanigfull and easier to scale. also tried to prepare code for possibility of changing who is player controlling. and also forgot to commit and also didnt work on this for months so i am not even sure. and changing godot to version 4.4 hence new .uid files and changes in tscn